+++ FeedLoader/FeedCache/LocalFeedLoader.swift
@@ -7,20 +7,6 @@
 
 import Foundation
 
-private final class FeedCachePolicy {
-    private init() {}
-    
-    private static let maxCacheAgeInDays = 7
-    private static let calendar = Calendar(identifier: .gregorian)
-    
-    static func validate(_ timestamp: Date, against date: Date) -> Bool {
-        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
-            return false
-        }
-        return date < maxCacheAge
-    }
-}
-
 final class LocalFeedLoader {
 
     private let store: FeedStore
